requirements.txt
streamlit
pandas
numpy
gspread
google-auth
python-dotenv
python-dateutil
supabase
